+++ CHANGELOG.md
@@ -3,14 +3,16 @@
 
 -------------------------------------------------------------------------------------------------------------
 
-## 5.3.10 (2020-07-16)
+## 5.3.10 (2020-07-17)
 
 ### 新特性
 * 【db   】       增加DbUtil.setReturnGeneratedKeyGlobal（issue#I1NM0K@Gitee）
 * 【core 】       增加DataSize和DataSizeUtil（issue#967@Github）
+* 【core 】       ImgUtil增加异常，避免空指针（issue#I1NKXG@Gitee）
 
 ### Bug修复
 * 【core   】     修复ZipUtil中finish位于循环内的问题（issue#961@Github）
+* 【core   】     修复CollUtil.page未越界检查的问题（issue#I1O2LR@Gitee）
 
 -------------------------------------------------------------------------------------------------------------
 ## 5.3.9 (2020-07-12)